server/requirements.txt
fastapi==0.115.0
uvicorn==0.30.1
motor==3.5.1
pymongo==4.7.2
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1